requirements.txt


# main
numpy==1.23.3
pyglet==1.5.15
pillow==9.2.0

## Alternatives
# pygame==2.1.2